llm-service/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
aiokafka>=0.11,<1
pydantic>=2.10,<3
pydantic-settings>=2.7,<3
structlog>=24.4,<26
prometheus-fastapi-instrumentator>=7.0,<8
httpx>=0.27,<1
google-genai>=1.0,<2
